Fix a type error and increase the warning level
I accidentally forgot to change it from the default of /W3 to the
typically preferable /W4, and the compiler didn't warn about a
mismatching wide/narrow string literal parameter.

There was such a bug. This fixes it, and increases the warning
level for the project to /W4. (The dialect remains set to the
classic MSVC dialect of C for compatibilty, i.e., to make it easier
to retarget this for testing on other, including older, systems.)